chore: extract test file watch as rsbuild plugin
